[igniteteams] feat: add exceptions handler

+++ igniteteams/src/storage/group/groupCreate.ts
@@ -1,10 +1,17 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { GROUP_COLLECTION } from "../storageConfig";
 import { groupsGetAll } from "./groupGetAll";
+import { AppError } from "@/utils/AppError";
 
 export async function groupCreate(newGroup: string) {
   try {
     const storedGroups = await groupsGetAll();
+
+    const isGroupAlreadyStored = storedGroups.includes(newGroup);
+
+    if (isGroupAlreadyStored) {
+      throw new AppError('Já existe um grupo cadastrado com esse nome.');
+    }
     const storage = JSON.stringify([...storedGroups, newGroup]);
 
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
